Refactor HomeScreen to display a list of countries
This commit introduces a `CountryModel` and updates the `HomeScreen`
to display a list of countries with their FIFA rankings using a `LazyColumn`.
The `NavController` is removed from `HomeScreen` as it's no longer directly
used for navigation within this screen.

Additionally, project configuration files and dependencies are updated,
including the Android Gradle Plugin version and the addition of the Coil
Compose library.