Enhance HomeScreen UI with country flags and details
This commit updates the `HomeScreen` to display a list of countries with their flags and FIFA rankings using `Card` and `Image` composables. The `rememberAsyncImagePainter` from Coil is used for image loading.

Additionally, the `INTERNET` permission has been added to the `AndroidManifest.xml` to allow network requests for fetching the flag images. A Kotlin error log file was also added.